feat(runloop) change treatment of rebuild_plugins_iterator failure
This change modifies Kong's behavior when an error happens while
rebuilding the router: previously, such an issue would cancel the
current request and return a 500.

This new implementation logs the error and uses the *existing* router to
handle the current request.

This is aligned with what we are doing with the plugins_iterator, but
more importantly, it provides more stability under heavy load, as
discussed #4194. It is also more aligned with the idea of async
rebuilds.

It also introduces `get_updated_router` in preparation for the
`async_rebuilds` config option.